src/main/java/parsing_json/ElementCollectionInitializer.java
package parsing_json;

import com.google.gson.Gson;
import com.google.gson.JsonElement;
import com.google.gson.JsonObject;
import com.google.gson.JsonParser;

import java.io.IOException;
import java.io.InputStream;
import java.io.InputStreamReader;
import java.io.Reader;
import java.util.ArrayList;

public class ElementCollectionInitializer {

    private ArrayList<String> list = new ArrayList<String>();
    String json = new Gson().toJson(list);

    public static ElementCollection generate() {

        return null;
    }

    public void testOnly() {
        JsonParser parser = new JsonParser();
        InputStream inputStream = getClass().getClassLoader().getResourceAsStream("periodic_table.json");
        Reader reader = new InputStreamReader(inputStream);
        JsonElement rootElement = parser.parse(reader);
        JsonObject rootObject = rootElement.getAsJsonObject();
    }


}
